chore: move details to main folder; update colors
