+++ lib/components/Widgets/DetailsItem/index.stories.tsx
