+++ lib/components/Widgets/DetailsItem/index.tsx
