+++ src/playground/exports.ts
@@ -1,3 +1,3 @@
-export * from "./Details"
+export * from "../../lib/components/Widgets/Details"
 export * from "./Pages/exports"
 export * from "./Widgets/exports"
